refactor: reduce code duplication and complexity with appwrite functions
