+++ server/appwrite/.env.example
@@ -9,5 +9,5 @@ APPWRITE_PASSWORD=your_password_here
 # Appwrite API Key (for function runtime operations)
 APPWRITE_API_KEY=your_api_key_here
 
-# NZTAB API Configuration (optional - defaults to https://api.tab.co.nz)
+# NZTAB API Configuration
 NZTAB_API_BASE_URL=https://api.tab.co.nz
