+++ server/appwrite/functions/daily-race-importer/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "daily-race-importer",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "src/main.js",
+  "dependencies": {
+    "node-appwrite": "^17.0.0"
+  }
+}
