+++ server/appwrite/functions/race-data-poller/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "race-data-poller", 
+  "version": "1.0.0",
+  "type": "module",
+  "main": "src/main.js",
+  "dependencies": {
+    "node-appwrite": "^17.0.0"
+  }
+}
