--- server/daily-race-importer/deploy-with-env.js
@@ -1,218 +1,0 @@
-import { readFileSync } from 'fs';
-import { spawn } from 'child_process';
-import { fileURLToPath } from 'url';
-import { dirname, join } from 'path';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-
-const FUNCTION_ID = '687c8695002e32d4272b';
-
-// Read and parse .env file
-function parseEnvFile(filePath) {
-  try {
-    const envContent = readFileSync(filePath, 'utf8');
-    const envVars = {};
-    
-    envContent.split('\n').forEach(line => {
-      line = line.trim();
-      if (line && !line.startsWith('#')) {
-        const [key, ...valueParts] = line.split('=');
-        if (key && valueParts.length > 0) {
-          const value = valueParts.join('=').trim();
-          // Remove quotes if present
-          envVars[key.trim()] = value.replace(/^["']|["']$/g, '');
-        }
-      }
-    });
-    
-    return envVars;
-  } catch (error) {
-    console.error(`❌ Error reading .env file: ${error.message}`);
-    process.exit(1);
-  }
-}
-
-// Execute shell command
-function executeCommand(command, args = []) {
-  return new Promise((resolve, reject) => {
-    console.log(`🔧 Executing: ${command} ${args.join(' ')}`);
-    
-    const child = spawn(command, args, {
-      stdio: 'inherit',
-      shell: true
-    });
-    
-    child.on('close', (code) => {
-      if (code === 0) {
-        resolve();
-      } else {
-        reject(new Error(`Command failed with exit code ${code}`));
-      }
-    });
-    
-    child.on('error', (error) => {
-      reject(error);
-    });
-  });
-}
-
-// Execute shell command and capture output
-function executeCommandWithOutput(command, args = []) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(command, args, {
-      stdio: ['inherit', 'pipe', 'pipe'],
-      shell: true
-    });
-    
-    let stdout = '';
-    let stderr = '';
-    
-    child.stdout.on('data', (data) => {
-      stdout += data.toString();
-    });
-    
-    child.stderr.on('data', (data) => {
-      stderr += data.toString();
-    });
-    
-    child.on('close', (code) => {
-      if (code === 0) {
-        resolve({ stdout, stderr });
-      } else {
-        reject(new Error(`Command failed with exit code ${code}: ${stderr}`));
-      }
-    });
-    
-    child.on('error', (error) => {
-      reject(error);
-    });
-  });
-}
-
-// Get existing variables for the function
-async function getExistingVariables() {
-  try {
-    const { stdout } = await executeCommandWithOutput('appwrite', [
-      'functions', 'list-variables',
-      '--function-id', FUNCTION_ID
-    ]);
-    
-    // Parse the CLI output (it's not JSON, it's a table format)
-    const variables = {};
-    const lines = stdout.split('\n');
-    let inVariablesSection = false;
-    
-    for (const line of lines) {
-      if (line.includes('variables')) {
-        inVariablesSection = true;
-        continue;
-      }
-      if (inVariablesSection && line.trim()) {
-        // Try to extract variable information from table format
-        const match = line.match(/(\w+)\s+/);
-        if (match) {
-          variables[match[1]] = true;
-        }
-      }
-    }
-    
-    return variables;
-  } catch (error) {
-    console.log(`⚠️ Could not fetch existing variables: ${error.message}`);
-    return {};
-  }
-}
-
-// Set environment variables for the function
-async function setEnvironmentVariables(envVars) {
-  console.log('⚙️ Setting environment variables...');
-  
-  const existingVars = await getExistingVariables();
-  let successCount = 0;
-  
-  for (const [key, value] of Object.entries(envVars)) {
-    try {
-      // Always try to create the variable (simpler approach)
-      await executeCommand('appwrite', [
-        'functions', 'create-variable',
-        '--function-id', FUNCTION_ID,
-        '--key', key,
-        '--value', value
-      ]);
-      console.log(`   ✅ Set ${key}`);
-      successCount++;
-    } catch (error) {
-      console.log(`   ⚠️ Variable ${key} might already exist or failed to create`);
-      // Variables might already exist, which is fine
-    }
-  }
-  
-  console.log(`✅ Environment variable setup completed (${successCount} processed)\n`);
-}
-
-async function deployWithEnvironmentVariables() {
-  try {
-    console.log('🚀 Starting deployment with environment variables...\n');
-    
-    // Parse environment variables from .env file
-    const envFilePath = join(__dirname, '.env');
-    console.log(`📄 Reading environment variables from: ${envFilePath}`);
-    const envVars = parseEnvFile(envFilePath);
-    
-    // Filter out deployment-only variables
-    const deploymentOnlyVars = ['APPWRITE_EMAIL', 'APPWRITE_PASSWORD'];
-    const functionEnvVars = Object.fromEntries(
-      Object.entries(envVars).filter(([key]) => !deploymentOnlyVars.includes(key))
-    );
-    
-    console.log(`✅ Found ${Object.keys(functionEnvVars).length} environment variables for function:`);
-    Object.keys(functionEnvVars).forEach(key => {
-      console.log(`   - ${key}`);
-    });
-    console.log('');
-    
-    // Deploy the function (no build step needed for JavaScript-only functions)
-    console.log('📦 Function ready (JavaScript source, no build required)\n');
-    
-    // Step 1: Deploy the function first to create it
-    console.log('🚀 Initial deployment to create function...');
-    await executeCommand('appwrite', ['push', 'functions']);
-    console.log('✅ Function created successfully\n');
-    
-    // Step 2: Set environment variables on the created function
-    await setEnvironmentVariables(functionEnvVars);
-    
-    // Step 3: Redeploy to apply environment variables
-    console.log('🔄 Redeploying to apply environment variables...');
-    await executeCommand('appwrite', ['push', 'functions', '--function-id', FUNCTION_ID]);
-    console.log('✅ Function redeployed with environment variables live\n');
-    
-    // Step 4: Check deployment status
-    console.log('🔄 Checking deployment status...');
-    try {
-      await executeCommand('appwrite', [
-        'functions', 'list-deployments',
-        '--function-id', FUNCTION_ID
-      ]);
-      console.log('✅ Deployment ready\n');
-    } catch (error) {
-      console.log('⚠️ Could not verify deployment status\n');
-    }
-    
-    console.log('🎉 Deployment completed successfully!');
-    console.log('📊 Summary:');
-    console.log(`   - Function created and deployed`);
-    console.log(`   - Environment variables set: ${Object.keys(functionEnvVars).length}`);
-    console.log('   - Function redeployed with live environment variables');
-    console.log('   - Ready for execution');
-    console.log('\n💡 Next steps:');
-    console.log('   - Run: npm run execute:this');
-    
-  } catch (error) {
-    console.error(`❌ Deployment failed: ${error.message}`);
-    process.exit(1);
-  }
-}
-
-deployWithEnvironmentVariables();
