--- server/daily-race-importer/functions/daily-race-importer/.prettierrc.json
@@ -1,6 +1,0 @@
-{
-  "trailingComma": "es5",
-  "tabWidth": 2,
-  "semi": true,
-  "singleQuote": true
-}
