--- server/daily-race-importer/functions/daily-race-importer/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "starter-template",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/main.js",
-  "type": "module",
-  "scripts": {
-    "format": "prettier --write ."
-  },
-  "dependencies": {
-    "node-appwrite": "^17.0.0"
-  },
-  "devDependencies": {
-    "prettier": "^3.2.5"
-  }
-}
