--- server/daily-race-importer/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "daily-race-importer",
-  "version": "1.0.0",
-  "description": "Daily race data import function for RaceDay application",
-  "main": "functions/daily-race-importer/src/main.js",
-  "type": "module",
-  "scripts": {
-    "dev": "node local-runner.js",
-    "deploy": "appwrite push functions",
-    "deploy:env": "node deploy-with-env.js",
-    "execute": "appwrite functions list && echo 'Enter function ID to execute:' && read FUNC_ID && appwrite functions create-execution --function-id $FUNC_ID",
-    "execute:this": "appwrite functions create-execution --function-id 687c8695002e32d4272b",
-    "login": "node login.js",
-    "logout": "appwrite logout",
-    "status": "appwrite functions list",
-    "format": "cd functions/daily-race-importer && npm run format"
-  },
-  "dependencies": {
-    "node-appwrite": "^17.0.0",
-    "dotenv": "^16.0.0"
-  },
-  "engines": {
-    "node": ">=22.0.0"
-  }
-}
