--- server/race-data-poller/.env.example
@@ -1,13 +1,0 @@
-# Appwrite Configuration
-APPWRITE_ENDPOINT=https://your-appwrite-instance.com/v1
-APPWRITE_PROJECT_ID=your_project_id_here
-
-# Appwrite User Credentials (for CLI deployment operations)
-APPWRITE_EMAIL=your_email@example.com
-APPWRITE_PASSWORD=your_password_here
-
-# Appwrite API Key (for function runtime operations)
-APPWRITE_API_KEY=your_api_key_here
-
-# NZTAB API Configuration (optional - defaults to https://api.tab.co.nz)
-NZTAB_API_BASE_URL=https://api.tab.co.nz
