--- server/race-data-poller/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "race-data-poller",
-  "version": "1.0.0",
-  "description": "Dynamic polling function for race data based on start time",
-  "main": "src/main.js",
-  "type": "module",
-  "scripts": {
-    "format": "prettier --write .",
-    "test": "node --test",
-    "dev": "node local-runner.js"
-  },
-  "dependencies": {
-    "node-appwrite": "^17.0.0"
-  },
-  "devDependencies": {
-    "prettier": "^3.2.5",
-    "dotenv": "^16.3.1"
-  }
-}
